Split c++ into a library/plugin
This splits c++ and qml into seperate parts, this is done to follow how
most of the components in ubuntu touch is built, it also allows other
applications to use the Matrix/uMatriks api (this idea can be expanded
on in the future). But this is mostly done due to call feature that I
have created requres this (oxide cannot handle qrc)

The way this works now, is that we now have a two "modules" one qml and
js that is not compiled/compressed this is the main "project" that
imports the libMatrix module.

This can now be started with qmlscene :)